changed double to BigDecimal

+++ src/main/java/edu/nbu/projectshop/tools/PriceCalculations.java
@@ -1,6 +1,7 @@
 package edu.nbu.projectshop.tools;
+import java.math.BigDecimal;
 import java.util.Date;
 public interface PriceCalculations {
-    double priceCalculation(double percentage, double deliveryPrice);
+    BigDecimal priceCalculation(BigDecimal percentage, BigDecimal deliveryPrice);
     boolean checkExpirationDate(Date currentDate, Date productionDate, int period);
 }
